Detect if the job was deleted during run
The jobs get removed now if the CRD is deleted. But if the CRD is
deleted while a backup is running the operator crashed. With this change
the operator is able to detect if the job was removed while running. The
job can also get deleted by other means. The operator won't crash any-
more now in these cases.