Unlock on sigterm
With this commit wrestic will perform a lock cleanup upon receiving a
sigterm. This should help to prevent most cases of stale locks that may
occur during normal OpenShift/Kubernetes operation. Also the prune job
will trigger an unlock all before running, as the operator ensures that
the prune is the only running job at that time.

Relates to: APPU-1313