Locking now considers the chronology of the jobs
If there are multiple jobs queued up the locker will now check when what
job was inserted into the queue. This fixes following issue:

Before when a backup took to long and a prune was scheduled, the prune
would start to run as soon as the backup was finished. But if another
backup got queued up before that there was a deadlock.

That deadlock was caused by the semaphores that didn't follow the
ordering of the jobs. So the queud up and already locked prune job would
wait for all backups to finish before running. But the backup that was
queued up after is waiting for that prune to finish. Both jobs waited
for each other to finish -> classic deadlock.