Fix scheduling redundant jobs
This fix is implemented on two levels:
First, when the Check handler tries to create a job
and it exists, it will not attemp to schedule the job again.
Second, it actually sets the Status.Started property on a
Check resource once the job is scheduled, which prevents
jobs to be created a second time in the first place.