Skip PreBackupPod integration test in certain cases
Because the implementation of PreBackupPods annoyingly relies on it's own K8s client instance,
which is independant of the instance that comes from the operator sdk,
it is not compatible with envtest. The code for PreBackupPods thous fails to run
when there is no actual k8s cluster available.
Because this is the case in our CD environment, these tests fail even though they might run fine
on a developers computer. Therefore we skip all testing of PreBackupPods when there is likely
no real k8s cluster available.
This is done by a simple heuristic: If KUBECONFIG is not set, and there is no ~/.kube/config file,
then there is likely no k8s cluster available.