Refactor STDOUT and STDERR log capture of restic
This prohibits a stack trace in the log output whenever
`restic` writes anything to stderr but exits with 0.

Previously, this could have lead to confusion when debugging a
problem with k8up – showing an error in the log when all is fine.

Also replaces deprecated logr.InfoLogger with logr.Logger.