fix: exclude backup pods from podlist
previously the executor would happily set `TARGET_PODS` for the backup
job to the pod of a previous running backup instead of a actually
running pod. This prevents `k8up.io/backupcommand` from working, as then
the backup job would actually check the old incarnation of itself
instead of the real target pod.

To do this we filter using the label we set on all backup pods.

Signed-off-by: huettner94 <huettner94@gmx.de>