Invert podlister k8s client dependency
This commit inverts the dependency of the `PodLister` to the k8s client.
Instead of generating one, the `PodLister` now accepts a k8s clent. This
ensures that the we can inject the correct client given the context.

Also the `PodLister` now uses the typed k8s client instead of the rest
client.

This solves a problem with the integration-tests.

Signed-off-by: Simon Beck <simon.beck@vshn.ch>