Harden RBAC for the operator
Previously the operator had the `pod/exe` permissions with the verb
`create`. Which is necessary for K8up to create and bind the namespaced
roles where the actual backups happen.

To harden the RBAC a bit we're now deploying a clusterRole which
contains the necessary RBAC rules. Also K8up is now allowed to bind
exactly this clusterRole.

While this doesn't mitigate all potential exploits in case the K8up
operator pod is accessed by a malicious party, it makes it harder for
the malicious party to abuse K8up's permissions.

Signed-off-by: Simon Beck <simon.beck@vshn.ch>