Improve tutorial manifests
The securityContext parameters should be available into the manifests.
Without it, testing the k8up on the MicroShift environment will raise
warnings:

    Warning: would violate PodSecurity "restricted:latest":
    seccompProfile (pod or container "mariadb" must set
    securityContext.seccompProfile.type to "RuntimeDefault" or "Localhost")

Also added the registry prefix to the images used in the tutorial, to
avoid complication when the runtime environment would not choose
docker.io as a default registry and removed namespace parameter from
secrets manifests, because other manifests in the tutorial does not have
such parameter (so if someone will use other namespace than 'default',
services/deployments will faile because of missing secrets).