Make e2e tests more reliable
`kubectl run --attach` proved to be rather unreliable. It often was not
able to attach to the newly created pod. There were various different
error messages.

By running the pod, waiting and then printing the logs, it improved the
reliability by a lot. Where it previously failed ever so often, I was
not able to have it fail once with the new function so far.

Signed-off-by: Simon Beck <simon.beck@vshn.ch>